testing and debugging the downloading scripts

+++ data_set_creation_script/download_PST900.py
@@ -1,20 +1,20 @@
+import gdown
 import os
-import dataset_tools as dtools
 import zipfile
 
 
-def download_PST900(output_path=None):
-    if output_path is None:
-        output_path = os.path.abspath("./real_data")
-    os.makedirs(output_path, exist_ok=True)
-
-    output = os.path.join(output_path, "PST900.zip")
-    dtools.download(dataset="PST900 RGB-T", dst_dir=output_path)
-
+def download_PST900(path=None):
+    if path is None:
+        path = os.path.abspath("./real_data")
+    os.makedirs(path, exist_ok=True)
+    url = f"https://drive.google.com/file/d/1hZeM-MvdUC_Btyok7mdF00RV-InbAadm/view"
+    output = os.path.join(path, "PST900.zip")
+    gdown.download(url, output, quiet=False, fuzzy=True)
+    if not os.path.exists(output):
+        raise Exception(f"Failed to download the dataset. from {url}")
     with zipfile.ZipFile(output, "r") as zip_ref:
-        zip_ref.extractall(output_path)
+        zip_ref.extractall(path)
     os.remove(output)
-    print("PST900 dataset downloaded and extracted successfully.")
 
 
 if __name__ == "__main__":
